Add files via upload
This article explains the SSL installation process for Windows-based servers when the CSR (Certificate Signing Request) and the corresponding Private Key were not generated on the server where the SSL needs to be installed.

To successfully install the certificate, you need to have the Private Key file. If the Private Key file is missing, you can search for it, or generate a new CSR and Private Key and reissue your SSL certificate.

When activating a certificate on SSLs.com, you have two options for submitting the CSR:

Using the “Auto-activate” option, where the CSR is generated in-browser
Manually entering the CSR code you generated on your hosting server
In this article, we’ll describe the most convenient way to install the certificate on a Windows-based server if you chose the first option or if the CSR was generated elsewhere, such as an online CSR generator.